+++ CHANGELOG.md
@@ -21,7 +21,8 @@
     - ``linux``, ``darwin``, ``cygwin``, ``emscripten``
 - show row numbers in completion pager if actual rows are larger than rendered rows
 - check read string length in builtin ``read`` command
-- check iterator invalidation of ``reply`` variable within builtin ``read`` and ``shctl info`` commands
+- check iterator invalidation of ``reply`` variable within some builtin commands
+    - ``printf``, ``read``, ``shctl info``
 
 #### Completion